DVC_Labs/Wine_Quality_Prediction/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
joblib==1.3.2
dvc==3.30.0